+++ vickyctl/src/locks/mod.rs
@@ -1,0 +1,6 @@
+mod http;
+mod tui;
+mod types;
+
+pub use tui::resolve_lock;
+pub use tui::show_locks;
